# perf(es/minifier): Use faster hash algorithm for DCE