--- a/crates/swc_ecma_transforms_optimization/Cargo.toml
+++ b/crates/swc_ecma_transforms_optimization/Cargo.toml
@@ -38,6 +38,7 @@ swc_ecma_transforms_base   = { version = "0.106.0", path = "../swc_ecma_transfor
 swc_ecma_transforms_macros = { version = "0.5.0", path = "../swc_ecma_transforms_macros" }
 swc_ecma_utils             = { version = "0.101.0", path = "../swc_ecma_utils" }
 swc_ecma_visit             = { version = "0.76.6", path = "../swc_ecma_visit" }
+swc_fast_graph             = { version = "0.15.4", path = "../swc_fast_graph" }
 tracing                    = "0.1.32"
 
 [dev-dependencies]